Define and enforce layer filename format.
This also splits layer keys into independent order and label
fields. The new file format looks like 001-some-label.yaml,
where both order and label are unique. This drops the benefit of
timestamps, originally described as it being easy to blindly
create, in favor of it being better to read and manipulate.
Seems like a good tradeoff to make.