Update layer config's "environment" to ordinary mapping
The original intention with the list of 1-item objects was so the
environment variables were ordered in case of dependencies, like
"PATH=$PATH:/foo/var" (not that we support interpolation right now).

However, it made for an awkward config format and awkward API,
especially in the Python Operator Framework. So we're going to a plain
dictionary for the environment, and we'll resolve any dependencies
using an explicit dependency check.

See https://github.com/canonical/pebble/issues/23