Return 200 for bulk operations unless the entire request is bad
This makes error handling on the client easier, because the "shape"
of the response is always the same if there's a non-200 status code --
it's not sometimes a list of path infos and sometimes an error dict.
If the response status is 200, you always get a list of infos
(including potential errors); if the status is not 200 the response
will always be of error type.