Fix timing issue with parser occasionally not seeing log message
This would happen once or twice in every 1000 runs of
TestMultipleServicesFollow locally (apparently more in CI). It happens
because sometimes when the it.Next() fires, the FormatWriter has only
written the timestamp and service name (first Write call), but not yet
called Write to write the actual log message.

Also update the test to ensure it tests Flush() behavior.