Add wait-change API endpoint and client function (#63)
This adds an endpoint to the Pebble API to wait for a change to be ready,
with an optional timeout. It's intended for use with the upcoming "exec"
feature in [PR](https://github.com/canonical/pebble/pull/61). The LXD
code has a similar feature, except changes are called "operations" in LXD,
and there's a "wait operation" API endpoint.

This is particularly important for "pebble exec", so we can wait for the
command to be finished without polling every short interval, and we can
return a response immediately when the command finishes executing.

In future we could also use this new endpoint in the `pebble start` and
`pebble stop` commands (and Python Operator Framework equivalents) to avoid
polling there too.