Make Exec return an opaque object with a Wait and Send* methods
This significantly simplifies usage of client.Exec, and cleans up
the messy Control goroutine and DataDone options. I'm not sure why LXD
had done it this way to begin with. This change also makes it more
similar to the Python client.

Also move getChangeWebsocket to changes.go where it belongs.