Revert to lxc exec / ssh behavior for detecting terminal mode
Per discussion at
https://github.com/canonical/pebble/pull/61#discussion_r719482813,
we wanted to make the default use-terminal mode based only on stdout,
but that's different to what both "lxc exec" and "ssh" do, so seems
wise to copy them. Additionally, it's problematic to implement,
because when allocating a PTY on the remote, you get a "master" file
descriptor that handles both stdin and stdout. You have to call
close() on the master FD to close stdin so that a command that reads
stdin (eg: "cat") stops, but you're closing the entire master FD so
that also closes stdout, causing the POLLNVAL errors we were seeing on
the stdout side, as well as the command output sometimes being cut
short.

We probably could address that by having a separate pipe for stdin and
specifying that in the API, or allocating two PTYs on the remote, or
similar, but it's complex and I don't think deviating from ssh or
"lxc exec" is a good idea.

In practice you very rarely want interactivity if you're piping stdin,
which is I guess why ssh (and lxc exec) do it the way they do.

When you try to pipe stdin and specify -t, here is what "lxc exec" and
ssh do:

* lxc exec: "echo foo | lxc exec <instance> -t -- cat" ... this fails
  with the error we were seeing here, exit code 129 and intermittent
  output for the same reason.
* ssh: "echo foo | ssh giftyweddings.com -t -- cat" ... because it's
  not going to work anyway, ssh explicitly prevents this, uses
  non-terminal mode anyway, and prints the following message on stderr:

$ echo foo | ssh giftyweddings.com -t -- cat
Pseudo-terminal will not be allocated because stdin is not a terminal.
foo

I think the stderr message is a good idea, because it tells you you've
done something wrong, so I've copied that here in "pebble exec".

This reverts commit d7ca5289e9d23b8f5d04c0bcc970f7939bce905c plus
adding the stderr msg if -t is specified and stdin is not a terminal.