Service auto-restart and backoffs (refactor as state machine) (#79)
This PR implements auto-restart and backoff functionality for Pebble services.
To achieve this, I've refactored the service handling into a state machine to
make it consistent and easy to add the new states like `backoff` -- most of
`internal/overlord/servstate/handlers.go` has been rewritten.

The PR adds the following new fields to the layer configuration under
services > name (documented in README.md and defaults shown below):

```
on-success: restart
on-failure: restart
backoff-delay: 500ms
backoff-factor: 2.0
backoff-limit: 30s
```

The PR also refactors how/when we track "active services", and means the logs
are saved across service stop/starts (and backoffs/restarts). So we're fixing
a few long-standing issues at once with the refactor.

The list of states is:

```
const (
    stateInitial serviceState = iota
    stateStarting
    stateRunning
    stateTerminating
    stateKilling
    stateStopped
    stateBackoff
)
```

When you start a service it starts at `initial`, then moves to `starting` for
the 1s okay-wait period, then transitions to `running`. When you stop a
service it sends SIGTERM then goes to `terminating` and then `stopped`
(via `killing` if SIGKILL is necessary). 

If a service is configured for auto-restart (the new default), it goes through
the `backoff` state, and there's a `backoffTime` to keep track of the current
backoff time. The new `Backoff*` fields in the service configuration specify
how the backoff delay will advance.

This PR also adds a new status for the `Current` field in the API/CLI:
`backoff`. Example CLI output:

```
$ pebble services
Service  Startup  Current
svc1     enabled  inactive
svc2     enabled  active
svc3     enabled  backoff
```

The PR also adds a `ServiceManager.SendSignal` function to send a signal to a
running service. This is only used for tests right now, but we're planning to
expose it via the API in another PR -- that will be easy to add.

The tests of restart and backoff currently use short timeouts, with some
`time.Sleep()` calls. This seems to work very reliably in my tests, but if it
becomes a problem we change change those to a wait/poll loop to wait for the
logs the service writes, and check the timestamp deltas are what we expect
(but it's a bit more work).