Simplify client.Logs "N" parameter and avoid pointer (#90)
This removes the pointer on the N *int (number of logs) parameter to be just a plain int. When calling Logs (non-follow), the default is the server-specified default (also changed to 30). When calling FollowLogs, the default is 0 (don't show any buffered logs).

This PR also changes the server API to have the same semantics (except that the default is "unset n" rather than "zero N"). This shouldn't affect any real users -- the pebble logs CLI command already sends through an explicit "n" value anyway. Any negative "n" values other than -1 are now an error (but this won't affect any clients).

Note: this is a breaking change in the Go client. However, it's a trivial fix, and at this point we probably have exactly one internal user of the Go client (the Pebble CLI itself), so it should be reasonable.