feat(#93): Wait for child processes
This feature implements waiting for child processes when stopping a
service via SIGTERM. This is done via waiting for all child processes
that are part of the process group of the main (service) process, i.e.
the leader process. The syscall used is Wait4 with a waitpid behavior.
Note that waiting for grandchildren is not supported with this approach.