Custom health checks: manager, check types, and on-check-failure handling (#85)
This implements custom health checks and the `on-check-failure` logic
per the [spec]
(https://docs.google.com/document/d/1d6-h3UAt2VPUSvlkVF30l8iuDW8raRNkHbp5M6NUo1A/edit).

Basically, you specify the list of custom health checks in a new
top-level configuration object named `checks` (see the `README.md`
documentation). They can be HTTP, TCP, or exec checks. Then in your
service configuration you specify any action you want to perform
`on-check-failure` for the various checks.

Here's an example configuration with a single HTTP-based check, and a
single service that will be restarted if the check fails (twice in a
row):

services:
    web-server:
        command: python3 webapp.py --port=8080
        on-check-failure:
            server-up: restart

checks:
    server-up:
        period: 5s
        failures: 2
        http:
            url: http://localhost:8080/is-up

It's a good amount of code, though at least half of it is tests. The
CLI and APIs are coming in
https://github.com/canonical/pebble/pull/86.

I'm pretty happy with how it turned out, particularly the checkers and
check manager. However, the wiring up of the check manager and the
service manager is a little odd, because they are mutually dependent
on one another: we have to tell the check manager about plan updates
(from the service manager), and we have to tell the service manager
about check failures (from the check manager). Open to suggestions on
that or anything else.