Don't log requests to /v1/health (similar to /v1/system-info) (#106)
Now that `/v1/health` is done via the standard daemon API endpoints
(albeit using an HTTP server) we're logging hits to it every 5s when
deployed on Juju K8s (the probes hit it every 5s). Add `/v1/health` to
the list of "skip log" URLs to avoid spamming the logs.