Fix panic when replan is called and there are no services to stop/start
When a replan was executed and there were no services to stop/start
(for example, when there are no startup: enabled services), the
POST /v1/services endpoint would panic due to the unhandled case.
Handle this case, which is a change with no tasks -- it needs to be a
real change because the endpoint returns a change ID (the client is
expecting a change ID or an error). It shouldn't be an error because
the replan is actually successful, there's just nothing to do.

Also add a test for this case.

This also removes newChange, which was only used in one place, and
less clear than just including it inline.

Fixes #94