Document that "pebble logs -f" only follows currently-running services
It would be quite an significant refactoring to have "pebble logs -f"
subscribe to new services being started and wire that through, so for
now at least, just document the current behaviour. It seems unlikely to
affect people in practice, as the service(s) you're debugging will
likely be running.