convert to timestamp detection
Makes sure to handle underscore/padded layout formats.
For dates like 1/_2/2006 that could result in either "12/31/2112" or "1/
1/2113" the number of whitespace separated fields is not fixed (1 vs.
2).  This handles that by replacing underscores in the layout with
spaces and determining the maximum number of space-separated fields in a
date.  It then parses incoming log timestamps, potentially with extra
content for e.g. cases where we collected 2 fields instead of the
necessary 1 as explained above, while checking for and allowing "extra
text" errors.  The "extra text" errors are used to determine the actual
timestamp prefix length and the trimmed writing then proceeds as normal.

Because of these and other ambiguities, it is possible that we may fail
to parse timestamps due to autodetecting the incorrect layout.  This is
now handled by counting how many parse failures occur
post-auto-detection.  When this count hits a certain threshold, the
current auto-detected layout is discarded and we revert back to layout
detection mode.