Make Pebble a child subreaper and reap zombie children (#101)
This PR sets Pebble as a Linux "child subreaper", meaning if it's not
running as PID 1, it'll get zombie (non-waited-on) children re-parented
to Pebble rather than whatever *is* running as PID 1, so that we can
reap them.

It also adds a "reaper" background task whose main purpose is to ensure
that zombie / non-waited-on child processes are reaped (i.e., waited on
and removed from the process table).

The reaper works by waiting for `SIGCHLD`, and on receipt of that
signal, reaps as many waiting children as are available. For PIDs that
Pebble has started directly using `os/exec`, instead of calling
`cmd.Wait` we now call `reaper.WaitCommand()` -- this registers the PID
with the reaper so it can notify the caller when it exited.

This is a bit klunky, but it avoids a race condition between `Wait4
(-1, ...)` waiting for all PIDs, and `cmd.Wait()` waiting for a single
PID -- they all go to the `Wait4(-1, ...)` now. `reaper.WaitCommand`
still needs to call `cmd.Wait()` internally after the reaper has reaped
the PID, so that `os/exec` cleans up goroutines and file descriptors
after itself (it expects `cmd.Wait` to return a "no child processes"
error).

This should fix issues like
https://github.com/canonical/pebble/issues/71.