Make write/push and make-dir chown intermediate dirs when setting user
Per user report, write/push and make-dir are not setting the user/group
on intermediate directories when the user and group is set. This can
lead to the file being unreachable when Pebble is running a workload
with a specific user:group.

Fix these API calls to chown the intermediate directories as needed.
Use the existing osutil.MkdirAllChown where appropriate.

Add tests for the above, including integration tests that require root.

Fixes #124