Make exec checks consistently not inherit parent environment variables (#131)
Currently it inherits them if you don't have any environment variables
set in the config, but does not inherit them if you have one or more
set. This is a quirk of the implementation, and how the
exec.Command.Env field works. Make them not inherited in both cases to
be consistent.

Fixes #127