Add PoC implementation of PushFile()
This is a rather small proof-of-concept implementation of the PushFile()
client API, to explore the sending of files using streams. There are
certain pain points which I would greatly appreciate feedback for, like:

 * The `fileUploader` struct (which I would refactor into a small set of
   functions, if possible)
 * Error handling/logging within the goroutine in charge of uploading
   the files.
 * Testing, taking into account that nearly no tests were written, is
   there a way of testing the multipart body from within the client
   test suite? I didn't find `cs.req` particularly useful...

Other than that, I would like feedback on the approach itself -- I might
be misusing streams here and maybe there's a simpler/safer/faster
method.