support changing service destinations live/dynamically
Previously, service destinations were stuffed irrevocably inside a
multiwirter for the service command when it was initialized+started.
This prevented new added layers from changing the log destinations for a
service after starting that service.  Now the destination list is
queried every time logs are generated and forwarded.  Notably, this is
just a static map lookup+iteration - it's not like we're dynamically
walking over the plan each time.  The function arg to forwarder is
mostly to facilitate clean testing - the alternative would be something
like passing in the log manager - which would be a pain to set up as a
test fixture.