layer validation for log forwarding, housekeeping
Each logging backend registers itself with the logstate package - this
auto-generates a mapping of backend type/name to constructors as well as
auto-registers validators for the plan logstate config for each backend.
All backend specific code is within a single place - in the single file
for each backend.  All a new backend needs to do is call the
logstate.ReisterLogBackend to add a new backend.