Remove test APIS (FakeDoRetry, DebugGet, DebugPost) from client package
Move them to export_test.go, which makes them only available for
tests, which is what they're intended and used for. Strictly speaking
this is a backwards-incompatible change, as these were exported, but it
seems extremely unlikely anyone is using these test utilities.

In our README we say: "We try to never change the underlying API itself
in a backwards-incompatible way, however, we may sometimes change the
Go client in backwards-incompatible ways." I think this is one of those
"sometimes".