Flesh out client package docs; small README improvements (#160)
This adds a brief package doc comment to the client package, as well as
doc comment to most types and methods in the client package.

Also make a few updates to the README. Removed two bullet points from
"Some details worth highlighting", as those are discussed in detail
below. Link to API docs instead of providing an inline example. Update
"TODO" section items.

Also add an example for the client package docs. It's not really run
(in the "too hard basket" for now), but at least it's type checked.