Add "current-since" field to services API and CLI (#115)
This PR adds a current-since field to the API, meaning the time at which the "current" field last changed, for example started (current changed from inactive to active), and so on. It's related to #104 (though in a previous iteration it was "start time", not "curren since").

Here's how it looks like the CLI:

    $ pebble services
    Service  Startup  Current  Since
    test2    enabled  active   today at 17:08 NZST

    $ pebble services --abs-time
    Service  Startup  Current  Since
    test2    enabled  active   2022-04-28T17:08:45+12:00

This PR also removes the "restarts" field which isn't being used and was internal-only. Per #104, we're going to use changes/tasks or events instead.