Use strutil from x-go library
Make Pebble depend on strutil from github.com/canonical/x-go and
remove the local copy.

This is the first step to maintain a master copy of code
common between Snapd, Chissel and Pebble in one central location.

At the time of writing, the strutil version in x-go is the most
up-to-date version including the latest changes from both Snapd
and Pebble, except the following two functions from the Pebble
strutil package which will be dropped:

1. strutil.MakeRandomString()
2. strutil.UUID()

This functionality is now provided by the x-go/randutil package,
as already used by Snapd.