Add optional arguments in Pebble layer schema
The goal is to introduce the concept of optional and
overrideable arguments for a Pebble service, but we
must ensure that the layering properties of Pebble
are preserved and respected. The proposed syntax re-uses
therefore the existing service “command” attribute:

    services:
	myservice:
	    command: myservice --db=/var/db [ --quiet ]

The markers [ and ] are suggested because these are
typical annotations for optional parameters and are
rarely seen naked because they are interpreted by shells.