Changes service status on success ignore.
When a service under Pebble's control exits with code 0 and has it's
OnSuccess service key configured to ignore Pebble set's the services
state to one of error.

This change configures this logic so that when a service layer specifies
ignoring a service that has exited with a code of 0 it is now placed
into an inactive state.