randutil: allow pseudo rng seeding by caller
The randutil package was promoted to the x-go library because it provides
helper functions needed by multiple projects.

Feedback from https://github.com/canonical/x-go/pull/17 highlighted the
need for seeding of the pseudo random dependent functions to be left to
the user of randutil (the crypto/rand functions are unaffected). The reason
is simply that the seeding requirements differ between use cases
(and projects), and it should be the responsibility of the caller to supply
the seed.

Example Usage:

prng := randutil.NewPseudoRand(nil)
:
tmpfile := prng.RandomString(12)
:
wait := prng.RandomDuration(time.Hour)

Apply the changes required to make Pebble use the new randutil api.