Refactor `main` package into `internals/cli`.
The `main` package is special in Go: it is reserved for the package
containing a `main()` function that serves as the entry point for the
program and it can't be imported.

This commit refactors all command-line functionality that will be
shared by other software into the `internals/cli` package, which will be
imported and used by the `main()` function to invoke the Pebble
command-line interface.