Refactor internals into an exportable package (#226)
This change guides Pebble towards an internal structure that enables us to create specific Pebble-based applications that implement custom commands and add functionality that is not relevant to the upstream Pebble repository. However, we don't want to create a full-blown fork of Pebble and have the maintenance burden of having to port changes from upstream to every fork; for this reason, this change does two things:

1. Rename the internal package to internals so that it can be imported by well-known Pebble-based applications.

2. Refactor all CLI functionality from the main package to the internals/cli package, so that other Pebble-based applications can import this package without having to reimplement this functionality.

Such Pebble-based applications will be developed in close proximity to Pebble itself, and this change is not a promise to support internal packages of Pebble externally. Only those privileged cases are honored.

This change does not intend to be exhaustive and instead aims to serve as the groundwork for achieving the described goal.