Allow non-root user to specify current user/group when running commands
This goes for services, exec, and health checks of type "exec".

This also fixes a long-standing bug in the user/group parameters for
"exec" health checks not working at all. This was due to not setting
cmd.SysProcAttr, so the "cmd.SysProcAttr.Credential =" line would panic
with a nil pointer error.

Also a drive-by fix to the README to update to the new terminology.
See also https://github.com/canonical/operator/pull/950

Fixes #166