Allow non-root to specify user/group to itself (#239)
When user and group (or user ID and group ID) options are given when running services and they're equal to the current user and group, avoid setting os/exec's syscall.Credential Uid and Gid, because they require root. In the case when they're explicitly asking to run the command as the current user, we can just run it directly, without requiring root.

This goes for services, exec (one-shot commands), and health checks of type "exec".

This PR also fixes a long-standing bug in the user/group options for "exec" health checks not working at all. This was due to not setting cmd.SysProcAttr, so the cmd.SysProcAttr.Credential = ... line would panic with a nil pointer error.

Also a drive-by fix to the README to update to the new terminology. See also canonical/operator#950.

Fixes #166.