cmd: add detection for containers and pid1
The system manager currently makes some assumptions about the environment
it is running in.

For example, it assumes that a shutdown program is available in userspace
and accessible with PATH configured appropriately.

internals/daemon/daemon.go:
:
cmd := exec.Command("shutdown", "-r", ...
:

This patch adds two detection mechanisms that will allow code to make
environment specific decisions in the future (not part of this patch):

- cmd.IsConfined() returns true if running inside a container runtime

- cmd.IsInit() returns true if the system manager was started as PID 1

In addition, the overlord code currently disables reboot failure detection
if the system manager is running as PID 1. However, this change is only
required for container runtimes, and not generically.

- Update the boot id workaround code to only apply for container runtimes.