daemon: add ability to use syscall based reboot
Reboot in the daemon is currently relying on userspace having a
shutdown command. This may not always be the case, as Pebble is
used in a wide variety of environments.

Add a SetSyscallReboot() function to the daemon to allow the daemon
to be initialised with a syscall based reboot implementation.