servstate: remove on-check-failure test races
This patch improves the following tests, all related to checker
functionality in servstate observed when running under stress:

TestOnCheckFailureRestartWhileRunning
TestOnCheckFailureRestartDuringBackoff
TestOnCheckFailureIgnore
TestOnCheckFailureShutdown

Since we are unit testing servstate, and not checkstate, this patch
adds code to more effectively control (mock) when the check failure
actions are delivered during the test. This removes the requirement
for tests to rely on finely tuned delays to check various service
life-cycle states.

The following changes are made:

- Round some of the delays to multiples of 100ms to highlight the
  fact these values are less important now. Delays intended to
  present infinite are replaced with 10s to ensure severe failures
  are observed if the test does not work as intended.

- Remove iteration code to wait for a checker action, and replace
  with channel sync code.

- Add some additional function comments to explain what the tests
  aim to confirm.