overlord: allow adding custom managers
This commit introduces StateEngine.SetTaskRunner and therefore relaxes
the constraint on the ordering of the managers in the StateEngine.
This change has made the inited field of Overlord obsolete: there is
no reason to guard against adding managers after Overlord.New anymore.
So the inited field has been removed too.

In turn, the relaxing of the ordering allows sister projects to add
managers to the Overlord. The Overlord of a given Daemon can now be
reached using a method of Daemon, so the sister project can use this
functionality on their own Daemon.