Use Command.WaitDelay to avoid Command.Wait blocking on stdin/out/err
Add WaitDelay to ensure cmd.Wait() returns in a reasonable timeframe if
the goroutines that cmd.Start() uses to copy Stdin/Stdout/Stderr are
blocked when copying due to a sub-subprocess holding onto them. Read
more details in these issues:

- https://github.com/golang/go/issues/23019
- https://github.com/golang/go/issues/50436

This isn't the original intent of kill-delay, but it seems reasonable
to reuse it in this context.

Fixes #149