Use reaper to run shutdown command if reaper is running
Currently the TestRebootHelper test fails intermittently with the
following failure:

daemon_test.go:773:
    c.Assert(err, IsNil)
... value *os.SyscallError = &os.SyscallError{Syscall:"wait", Err:0xa} ("wait: no child processes")

This is because we're running the reaper, but systemdModeReboot doesn't
use or know about the reaper. So avoid starting/stopping the reaper
in the setup and teardown; daemon.New and Daemon.Stop -- via the
service manager) starts/stop it in any case. And set the FakeCommand
withReaper flag to false here as the reaper isn't running there now.

Fixes #186