feat: add "health" command
This commit adds a new "health" command to pebble. The health command
queries the health of configured pebble checks. It outputs "healthy" and
returns an exit code of 0 if all the specified checks are healthy.
Otherwise, it outputs "unhealthy" and returns with an exit code of 1.

The "health" command accepts two options with arguments:

  - The --level option takes an argument, either "alive" or "ready". It
    queries the health of only those checks with the same level.
    Note that, ready implies alive. Thus, if the ready checks are
    healthy but one of the alive checks are not, the overall output will
    be "unhealthy".

  - The --format option takes an argument, either "text" or "json". It
    controls how the output will look like. The default option is
    "text".

Usage:

  pebble health [health-OPTIONS] [<check>...]

This command does not add any new logic to pebble, as it relies entirely
on the already existing implementation of the pebble health API.

This commit includes a few tests too for the relevant changes.